requirements.txt
fastapi[all]~=0.111.0
uvicorn[standard]~=0.30.1
sqlmodel~=0.0.19
alembic~=1.13.1
psycopg2-binary
psycopg2-pool
passlib[bcrypt]~=1.7.4
python-dotenv~=1.0.1
python-jose[cryptography]
GeoAlchemy2~=0.15.1
fastapi-pagination
SQLAlchemy~=2.0.31
sqlalchemy_json
mailchimp_transactional
python-multipart
boto3~=1.34.132
paho.mqtt
stripe~=10.0.0
timezonefinder~=6.5.2
asyncio~=3.4.3
twilio~=9.2.1

python-dateutil~=2.9.0.post0
botocore~=1.34.132
pydantic~=2.7.4
starlette~=0.37.2
paho-mqtt~=2.1.0
requests~=2.32.3
pyotp
pandas
phonenumbers
apscheduler
pillow
